rename timepars to time
