+++ starsim/time.py
